+++ packages/block-library/src/navigation/edit/utils.js
@@ -2,6 +2,7 @@
  * External dependencies
  */
 import clsx from 'clsx';
+import { paramCase as kebabCase } from 'change-case';
 
 function getComputedStyle( node ) {
 	return node.ownerDocument.defaultView.getComputedStyle( node );
@@ -110,3 +111,72 @@ export function getNavigationChildBlockProps( innerBlocksColors ) {
 		},
 	};
 }
+
+/**
+ * Parses a template part ID into theme and slug components.
+ *
+ * This is the inverse of createTemplatePartId from:
+ * packages/block-library/src/template-part/edit/utils/create-template-part-id.js
+ *
+ * @param {string} templatePartId Template part ID in format "theme//slug".
+ * @return {{theme: string, slug: string}|null} Parsed components or null if invalid.
+ */
+export const parseTemplatePartId = ( templatePartId ) => {
+	if ( ! templatePartId || typeof templatePartId !== 'string' ) {
+		return null;
+	}
+
+	const parts = templatePartId.split( '//' );
+	if ( parts.length !== 2 ) {
+		return null;
+	}
+
+	return {
+		theme: parts[ 0 ],
+		slug: parts[ 1 ],
+	};
+};
+
+/**
+ * Return a unique template part title based on
+ * the given title and existing template parts.
+ *
+ * This implementation is copied from:
+ * packages/fields/src/components/create-template-part-modal/utils.js
+ *
+ * @param {string} title         The original template part title.
+ * @param {Object} templateParts The array of template part entities.
+ * @return {string} A unique template part title.
+ */
+export const getUniqueTemplatePartTitle = ( title, templateParts ) => {
+	const lowercaseTitle = title.toLowerCase();
+	const existingTitles = templateParts.map( ( templatePart ) =>
+		templatePart.title.rendered.toLowerCase()
+	);
+
+	if ( ! existingTitles.includes( lowercaseTitle ) ) {
+		return title;
+	}
+
+	let suffix = 2;
+	while ( existingTitles.includes( `${ lowercaseTitle } ${ suffix }` ) ) {
+		suffix++;
+	}
+
+	return `${ title } ${ suffix }`;
+};
+
+/**
+ * Get a valid slug for a template part.
+ * Currently template parts only allow latin chars.
+ * The fallback slug will receive suffix by default.
+ *
+ * This implementation is copied from:
+ * packages/fields/src/components/create-template-part-modal/utils.js
+ *
+ * @param {string} title The template part title.
+ * @return {string} A valid template part slug.
+ */
+export const getCleanTemplatePartSlug = ( title ) => {
+	return kebabCase( title ).replace( /[^\w-]+/g, '' ) || 'wp-custom-part';
+};
